auto-formatting by prettier-plugin-solidity

+++ src/DssDeployPauseProxyActionsAddon.sol
@@ -1,16 +1,32 @@
-
 pragma solidity >=0.5.12;
 
 interface PauseLike {
-    function plot(address, bytes32, bytes calldata, uint) external;
-    function exec(address, bytes32, bytes calldata, uint) external;
+    function plot(
+        address,
+        bytes32,
+        bytes calldata,
+        uint256
+    ) external;
+
+    function exec(
+        address,
+        bytes32,
+        bytes calldata,
+        uint256
+    ) external;
 }
 
 contract DssDeployPauseProxyActionsAddon {
-
-    function deny(address pause, address actions, address who, address to) external {
+    function deny(
+        address pause,
+        address actions,
+        address who,
+        address to
+    ) external {
         bytes32 tag;
-        assembly { tag := extcodehash(actions) }
+        assembly {
+            tag := extcodehash(actions)
+        }
         PauseLike(pause).plot(
             address(actions),
             tag,
