+++ src/base.sol
@@ -2,13 +2,14 @@ pragma solidity ^0.5.0;
 
 import "./IERC20.sol";
 
-
 interface IAdapter {
-    function calc(address gem, uint acc, uint factor) external view returns (uint);
+    function calc(
+        address gem,
+        uint256 acc,
+        uint256 factor
+    ) external view returns (uint256);
 }
 
 interface IGemForRewardChecker {
     function check(address gem) external view returns (bool);
 }
-
-
