+++ src/priceProvider.sol
@@ -1,6 +1,5 @@
 pragma solidity ^0.5.12;
 
-
 import "./IERC20.sol";
 import "./safeMath.sol";
 import "./safeERC20.sol";
@@ -13,44 +12,47 @@ interface RegistryLike {
     function list() external view returns (bytes32[] memory);
 }
 
-
-
 contract PriceProvider {
-
     using SafeMath for uint256;
     using SafeERC20 for IERC20;
 
-    mapping(address => uint) public rewards;
+    mapping(address => uint256) public rewards;
     RegistryLike public registry;
     SpotLike public spot;
     address public gov;
     address public owner;
-    uint public nextUpdate;
-    uint public updatePeriod;
-    uint public rewardPerPeriod;
-    uint public distributedReward;
-    uint public rewardToDistribute;
-    uint public rewardTime;
+    uint256 public nextUpdate;
+    uint256 public updatePeriod;
+    uint256 public rewardPerPeriod;
+    uint256 public distributedReward;
+    uint256 public rewardToDistribute;
+    uint256 public rewardTime;
 
     event RewardPaid(address indexed user, uint256 reward);
 
-    constructor () public {
+    constructor() public {
         owner = msg.sender;
     }
 
-    function setup(address _gov, address _spot, address _registry, uint _updatePeriod, uint _rewardTime) public {
+    function setup(
+        address _gov,
+        address _spot,
+        address _registry,
+        uint256 _updatePeriod,
+        uint256 _rewardTime
+    ) public {
         require(owner == msg.sender, "auth-error");
 
         require(_gov != address(0), "gov is null");
         require(_spot != address(0), "spot is null");
         require(_updatePeriod != 0, "updatePeriod is zero");
         require(_registry != address(0), "registry is null");
-        require(_rewardTime > _updatePeriod*10, "rewardTime vs updatePeriod inconsistence");
+        require(_rewardTime > _updatePeriod * 10, "rewardTime vs updatePeriod inconsistence");
 
         rewardToDistribute = IERC20(_gov).balanceOf(address(this));
         require(rewardToDistribute > 0, "no reward to distribute");
 
-        uint chunks = _rewardTime.div(_updatePeriod);
+        uint256 chunks = _rewardTime.div(_updatePeriod);
 
         registry = RegistryLike(_registry);
         spot = SpotLike(_spot);
@@ -61,9 +63,8 @@ contract PriceProvider {
         require(rewardPerPeriod > 0, "rewardPerPeriod is zero");
     }
 
-
-    function getReward() public returns (uint){
-        uint acc = rewards[msg.sender];
+    function getReward() public returns (uint256) {
+        uint256 acc = rewards[msg.sender];
         if (acc > 0) {
             distributedReward = distributedReward.add(acc);
             IERC20(gov).safeTransfer(msg.sender, acc);
@@ -74,9 +75,8 @@ contract PriceProvider {
     }
 
     function poke() public {
-
         bytes32[] memory ilks = registry.list();
-        for (uint i=0; i<ilks.length; i++) {
+        for (uint256 i = 0; i < ilks.length; i++) {
             spot.poke(ilks[i]);
         }
